pyproject.toml
[project]
name = "faa-poster"
version = "0.1.0"
description = "PosterStack — AI poster generation with gstack + evolver"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "FAA Team" }]
dependencies = [
    "pydantic>=2.0",
    "jinja2>=3.1",
    "pillow>=10.0",
    "playwright>=1.40",
    "requests>=2.31",
    "rich>=13.0",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
]

[project.scripts]
faa-poster = "faa_poster.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
